map helpers file removed, functions redistributed
